pyproject.toml
[tool.poetry]
name = "hero_camp"
description = ""
version = "1.0.0"
authors = ["Semen Syrovatskiy <mail@syrovatskiy.tk>"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.7.7"
asyncpg = "0.25.0"
bcrypt = "3.2.2"
fastapi = "0.78.0"
passlib = "1.7.4"
psycopg2-binary = "2.9.3"
pyjwt = "2.4.0"
SQLAlchemy = "1.4.36"
uvicorn = "0.17.6"

[tool.poetry.dev-dependencies]
async-asgi-testclient = "1.4.10"
black = "22.3.0"
isort = "5.10.1"
mypy = "0.950"
pytest = "7.1.2"
pytest-asyncio = "0.18.3"

[tool.poetry.scripts]
start = "hero_camp.asgi:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
exclude = [".venv", "migrations", "tests"]
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.isort]
py_version=310
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"
